modules/crop_quality_estimator/requirements.txt
# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Computer vision
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Numerics
numpy>=1.24.0

# UI (optional)
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0
